Legend key for the pie glyphs

Controls the aesthetics of the legend entries for the pie glyphs

Usage

draw_key_pie(data, params, size)

Arguments

data

A single row data frame containing the scaled aesthetics to display in this key

params

A list of additional parameters supplied to the geom.

size

Width and height of key in mm.

Value

A grid grob

See also